Killed off some more warnings. Changed order of parameters int,timeT to timeT, int.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@1343 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/rg21io.h
@@ -107,7 +107,7 @@ protected:
     EventIdMap m_indicationsExtant;
 
     bool m_inGroup;
-    unsigned int m_groupId;
+    long m_groupId;
     std::string m_groupType;
     Rosegarden::timeT m_groupStartTime;
     int m_groupTupledLength;
